requirements.txt
streamlit
PyGithub
feedparser
pandas
plotly
watchdog
openai
beautifulsoup4
